// app/Models/Reg/Parents/reg-parents.ts
export class regParents {
        id: number;
        idNum: string;
        fatherName: string;  // Full Father Name
        firstName: string;
        secondName: string;
        familyName: string;
        fullName: string;
        // fatherFirstName: string;
        // fatherSecondName: string;
        // fatherFamilyName: string;
        motherName: string;
        religionId: number;
        religionName: string;
        nationalityId: number;
        nationalityName: string;
        cityId: number;
        cityName: string;
        locality1: string;
        locality2: string;
        address: string;
        street: string;
        buildingNo: number;
        fatherEducationId: number;
        fatherEducationName: string;
        fatherSpec: string;
        matherEducationId: number;
        matherEducationName: string;
        matherSpec: string;
        parentRelationId: number;
        parentRelationName: string;
        parentName: string;
        parentWork: string;
        familySize: number;
        familyIncome: number;
        familyAssistance: string;
        refugeeCardNo: number;
        tel: string;
        fatherMobile: null;
        motherMobile: string;
        smsParent: string;
        smsMobile: string;
        parentEmail: string;
        parentFace: string;
        note: string;
        parentTotalPrice: number;
        parentTotalDescount: number;
        parentNetTotalAmt: number;
        FparentName:string;
}